+++ src/app/UILayout.tsx
@@ -1,12 +1,12 @@
 import { Box, Flex } from '@chakra-ui/react'
 import { type FC, type ReactNode } from 'react'
 
-import LayoutHeader from '~/components/LayoutHeader'
+import Header from '~/components/Header'
 
-const MainLayout: FC<{ children: ReactNode }> = ({ children }) => {
+const UILayout: FC<{ children: ReactNode }> = ({ children }) => {
 	return (
 		<Flex height="100%" direction="column">
-			<LayoutHeader shadow="lg" width="full" />
+			<Header shadow="lg" width="full" />
 			<Box as="main" flex="1" mt={6} pb={8}>
 				{children}
 			</Box>
@@ -14,4 +14,4 @@ const MainLayout: FC<{ children: ReactNode }> = ({ children }) => {
 	)
 }
 
-export default MainLayout
+export default UILayout
